Add some safety measures to the lexer (#9)
* Add some safety measures to the lexer
* added some asserts to the lexing function
* switched from UInt32 to Int32 for eols
* introduced `newlinechar` with explicit support for single byte chars
* dropped the half-baked support for escapes outside of quoted field
* Actually, don't restrict row numbers to 32 bit
* More informative errors for multibyte control characters

+++ src/exceptions.jl
@@ -5,9 +5,9 @@ Base.showerror(io::IO, e::FatalLexingError) = print(io, e.msg)
 #       and/or the quote character.
 struct NoValidRowsInBufferError <: FatalLexingError
     msg::String
-    buffersize::UInt32
+    buffersize::Int
 
-    function NoValidRowsInBufferError(buffersize::UInt32)
+    function NoValidRowsInBufferError(buffersize)
         return new(
             string(
                 "CSV parse job failed on lexing newlines. There was no linebreak (outside a quoted string field) ",
