+++ src/parser_doublebuffer.jl
@@ -1,6 +1,6 @@
 # TODO: start processing the "next" buffer as soon as it is ready (now we wait for the previous one to be fully processed)
 function read_and_lex_task!(parsing_queue::Channel{T}, lexer_state::LexerState{B}, parsing_ctx::ParsingContext, parsing_ctx_next::ParsingContext, consume_ctx::AbstractConsumeContext, options::Parsers.Options) where {T,B}
-    row_num = UInt32(1)
+    row_num = 1
     parsers_should_use_current_context = true
     done = lexer_state.done
     @inbounds while true
@@ -10,19 +10,19 @@ function read_and_lex_task!(parsing_queue::Channel{T}, lexer_state::LexerState{B
         # Set the expected number of parsing tasks
         setup_tasks!(consume_ctx, parsing_ctx, ntasks)
         # Send task definitions (segment of `eols` to process) to the queue
-        task_start = UInt32(1)
-        task_num = UInt32(1)
+        task_start = Int32(1)
+        task_num = Int32(1)
         for task in Iterators.partition(eachindex(parsing_ctx.eols), task_size)
-            task_end = UInt32(last(task))
+            task_end = Int32(last(task))
             put!(parsing_queue, (task_start, task_end, row_num, task_num, parsers_should_use_current_context))
-            row_num += task_end - task_start
+            row_num += Int(task_end - task_start)
             task_start = task_end
-            task_num += UInt32(1)
+            task_num += Int32(1)
         end
         # Start parsing _next_ chunk of input
         if !done
             last_newline_at = lexer_state.last_newline_at
-            unsafe_copyto!(parsing_ctx_next.bytes, last_newline_at, parsing_ctx.bytes, last_newline_at, UInt32(length(parsing_ctx.bytes)) - last_newline_at + 1)
+            unsafe_copyto!(parsing_ctx_next.bytes, last_newline_at, parsing_ctx.bytes, last_newline_at, Int32(length(parsing_ctx.bytes)) - last_newline_at + 1)
             read_and_lex!(lexer_state, parsing_ctx_next, options)
         end
         # Wait for parsers to finish processing current chunk
@@ -61,13 +61,13 @@ function process_and_consume_task(parsing_queue::Channel{T}, result_buffers::Vec
 end
 
 function _parse_file_doublebuffer(lexer_state::LexerState, parsing_ctx::ParsingContext, consume_ctx::AbstractConsumeContext, options::Parsers.Options, ::Val{M}) where {M}
-    parsing_queue = Channel{Tuple{UInt32,UInt32,UInt32,UInt32,Bool}}(Inf)
+    parsing_queue = Channel{Tuple{Int32,Int32,Int,Int32,Bool}}(Inf)
     result_buffers = TaskResultBuffer{M}[TaskResultBuffer{M}(id, parsing_ctx.schema, cld(length(parsing_ctx.eols), parsing_ctx.maxtasks)) for id in 1:parsing_ctx.nresults]
     parsing_ctx_next = ParsingContext(
         parsing_ctx.schema,
         parsing_ctx.header,
         Vector{UInt8}(undef, length(parsing_ctx.bytes)),
-        BufferedVector{UInt32}(Vector{UInt32}(undef, length(parsing_ctx.eols)), 0),
+        BufferedVector{Int32}(Vector{Int32}(undef, length(parsing_ctx.eols)), 0),
         parsing_ctx.limit,
         parsing_ctx.nworkers,
         parsing_ctx.maxtasks,
@@ -95,7 +95,7 @@ function _parse_file_doublebuffer(lexer_state::LexerState, parsing_ctx::ParsingC
     end
     # Cleanup
     for _ in 1:parsing_ctx.nworkers
-        put!(parsing_queue, (UInt32(0), UInt32(0), UInt32(0), UInt32(0), true))
+        put!(parsing_queue, (Int32(0), Int32(0), 0, Int32(0), true))
     end
     foreach(wait, parser_tasks)
     close(parsing_queue)
