+++ src/parser_serial.jl
@@ -1,17 +1,17 @@
 function _parse_file_serial(lexer_state::LexerState, parsing_ctx::ParsingContext, consume_ctx::AbstractConsumeContext, options::Parsers.Options, ::Val{M}) where {M}
-    row_num = UInt32(1)
+    row_num = 1
     result_buf = TaskResultBuffer{M}(0, parsing_ctx.schema, cld(length(parsing_ctx.eols), parsing_ctx.maxtasks))
     try
         while true
             limit_eols!(parsing_ctx, row_num) && break
             task_size = estimate_task_size(parsing_ctx)
-            task_start = UInt32(1)
+            task_start = Int32(1)
             for task in Iterators.partition(eachindex(parsing_ctx.eols), task_size)
                 setup_tasks!(consume_ctx, parsing_ctx, 1)
-                task_end = UInt32(last(task))
+                task_end = Int32(last(task))
                 _parse_rows_forloop!(result_buf, view(parsing_ctx.eols, task_start:task_end), parsing_ctx.bytes, parsing_ctx.schema, options, parsing_ctx.comment)
                 consume!(consume_ctx, parsing_ctx, result_buf, row_num, task_start)
-                row_num += task_end - task_start
+                row_num += Int(task_end - task_start)
                 task_start = task_end
                 task_done!(consume_ctx, parsing_ctx, result_buf)
                 sync_tasks(consume_ctx, parsing_ctx, 1)
